requirements.txt
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
decorator==5.1.1
executing==0.8.3
Flask==2.0.3
idna==3.3
iniconfig==1.1.1
ipython==8.1.1
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
MouseInfo==0.1.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pluggy==1.0.0
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
PyAutoGUI==0.9.53
pycodestyle==2.8.0
PyGetWindow==0.0.9
Pygments==2.11.2
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
pytest==7.0.1
python-dotenv==0.19.2
python3-xlib==0.15
pytweening==1.0.4
pywhatkit==5.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
stack-data==0.2.0
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.3
wikipedia==1.4.0
